refactor input text
