+++ src/app/components/pages/createdForm/createdForm.module.css
@@ -3,3 +3,37 @@
   flex-direction: column;
   align-items: center;
 }
+
+.input__form {
+  width: 95%;
+  padding: 5px 10px;
+  border-radius: 10px;
+  border: none;
+}
+
+.form__content {
+  display: flex;
+  justify-content: center;
+}
+
+.form__container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+  background-color: #dcdcdc;
+  border-radius: 10px;
+  padding: 20px;
+  width: 50%;
+}
+
+.input__container {
+  width: 100%;
+}
+
+.form__title {
+  text-align: center;
+  margin-bottom: 15px;
+  font-size: 24px;
+  font-weight: 500;
+}
