+++ src/app/ui/atoms/input/input.types.ts
@@ -3,4 +3,5 @@ export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
   placeholder: string;
   className: string;
   value: string;
+  type?: string;
 }
