--- src/app/ui/atoms/inputText/index.ts
@@ -1,1 +1,0 @@
-export { default as InputText } from './inputText';
